FIX: Ensure auras in AurasImportantList always shown
PROBLEM: The "Important Aura" list gives the user the impression of a Brute Force Whitelist, where auras added to the list are always shown, regardless of other conditions.

SOLUTION: Previously, the visibility of auras included in the AurasImportantList was dependent on several other conditions. This change modifies the 'toggle' variable assignment to ensure that if an aura is in the AurasImportantList, 'toggle' will evaluate to true and display the aura.